Class updates and Hand class
Made updates to DeckOfCards and Card class. Created a Hand class and a
HoldemHand class that inherits from Hand.
Need to make a FiveCardHand class that holds "potential" hands made from
cards on the game board and cards in the players' hands.
Will use these Hand classes in a different class that determines the
winning hand out of a list of hands given to the class.