Project Structure Changes
Worked a bit on the BestHand class.

Reorganized the file structure so that the project isn't in the defaultpackage, but is in a package called texasholdem. Also renamed the project to TexasHoldem.

Because of the above changes, I had to move around the resource files and modify the strings that direct the program to get the files.